+++ tests/test_fields.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 from django.test.utils import override_settings
 
-from pgcrypto_fields import fields
+from pgcrypto_fields import aggregates, fields
 from .factories import EncryptedTextFieldModelFactory
 from .models import EncryptedTextFieldModel
 
@@ -46,3 +46,14 @@ class TestEncryptedTextFieldModel(TestCase):
 
         instance = EncryptedTextFieldModel.objects.get()
         self.assertIsInstance(instance.encrypted_value, memoryview)
+
+    def test_value(self):
+        """Assert we can get back the decrypted value."""
+        expected = 'bonjour'
+        EncryptedTextFieldModelFactory.create(encrypted_value=expected)
+
+        queryset = EncryptedTextFieldModel.objects.annotate(
+            aggregates.Decrypt('encrypted_value'),
+        )
+        instance = queryset.get()
+        self.assertEqual(instance.encrypted_value__decrypt, expected)