+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 ## Upcoming
+
+* Add decryption through an aggregate class.
+* Add encryption when inserting data to the database.